Move structs and functions related to command arguments parsing into new separate file
